added the wining pct to players

+++ db/migrate/20131024234648_add_winning_pct_to_registered_players.rb
@@ -1,5 +1,5 @@
 class AddWinningPctToRegisteredPlayers < ActiveRecord::Migration
   def change
-    add_column :registered_players, :winning_pct, :double
+    add_column :registered_players, :winning_pct, :float
   end
 end
